fix: configure: check OCaml packages
The usage of the macro `AC_OCAML_CHECK_PKG` is as follows:
```
AC_OCAML_CHECK_PKG([pkgname], [alternative names...])
```
and the second argument can be omitted.

According to the implementation of the macro (found in `m4/ocaml.m4`),
if the several packages are given to the first argument, a loop will
iterate on them and exit at the first package found. This is not
what we want.

Therefore, we need to call `AC_OCAML_CHECK_PKG` for every packages.

Moreover, we also need to check that the packages were indeed found
and exit with a error if not.

Finally, checking for package `lwt` is not enough. We actually need to
check for the package `lwt_ppx` (this implies that `lwt` is present).
Installing `lwt` with opam is not enough.